Move NamedTuples to abi_tuples folder in structures.
