+++ READMEs/dispenser-flow.md
@@ -45,7 +45,7 @@ Please refer to [datatokens-flow](datatokens-flow.md) and complete the following
 In the same python console:
 ```python
 from ocean_lib.models.dispenser import Dispenser
-from ocean_lib.models.models_structures import DispenserData
+from ocean_lib.structures.abi_tuples import DispenserData
 from ocean_lib.ocean.util import get_address_of_type
 
 # Get the dispenser
