+++ ocean_lib/models/bfactory.py
@@ -6,7 +6,7 @@ from typing import Union
 
 from enforce_typing import enforce_types
 
-from ocean_lib.models.models_structures import BPoolData
+from ocean_lib.structures.abi_tuples import BPoolData
 from ocean_lib.web3_internal.contract_base import ContractBase
 from ocean_lib.web3_internal.wallet import Wallet
