+++ ocean_lib/models/bpool.py
@@ -8,7 +8,7 @@ from enforce_typing import enforce_types
 
 from ocean_lib.models import balancer_constants
 from ocean_lib.models.btoken import BTokenBase
-from ocean_lib.models.models_structures import BPoolInitialized
+from ocean_lib.structures.abi_tuples import BPoolInitialized
 from ocean_lib.web3_internal.wallet import Wallet
