+++ ocean_lib/models/factory_router.py
@@ -7,7 +7,7 @@ from typing import List, Union
 from enforce_typing import enforce_types
 
 from ocean_lib.models.bfactory import BFactory
-from ocean_lib.models.models_structures import Operations, Stakes
+from ocean_lib.structures.abi_tuples import Operations, Stakes
 from ocean_lib.web3_internal.wallet import Wallet
