+++ ocean_lib/models/test/test_erc20_token.py
@@ -11,11 +11,7 @@ from web3.main import Web3
 from ocean_lib.models.erc20_token import ERC20Token, RolesERC20
 from ocean_lib.models.erc721_factory import ERC721FactoryContract
 from ocean_lib.models.erc721_nft import ERC721NFT
-from ocean_lib.models.models_structures import (
-    CreateErc20Data,
-    ProviderFees,
-    ConsumeFees,
-)
+from ocean_lib.structures.abi_tuples import ConsumeFees, CreateErc20Data, ProviderFees
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
 from ocean_lib.web3_internal.utils import split_signature
