+++ ocean_lib/models/test/test_fixed_rate_exchange.py
@@ -13,7 +13,7 @@ from ocean_lib.models.fixed_rate_exchange import (
     FixedRateExchangeDetails,
     FixedRateExchangeFeesInfo,
 )
-from ocean_lib.models.models_structures import FixedData
+from ocean_lib.structures.abi_tuples import FixedData
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
 from tests.resources.helper_functions import deploy_erc721_erc20, get_address_of_type
