+++ ocean_lib/ocean/test/test_ocean_assets.py
@@ -17,12 +17,12 @@ from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 from ocean_lib.exceptions import AquariusError, ContractNotFound, InsufficientBalance
 from ocean_lib.models.erc721_factory import ERC721FactoryContract
 from ocean_lib.models.erc721_nft import ERC721NFT
-from ocean_lib.models.models_structures import (
+from ocean_lib.services.service import Service
+from ocean_lib.structures.abi_tuples import (
     ConsumeFees,
     CreateErc20Data,
     CreateERC721Data,
 )
-from ocean_lib.services.service import Service
 from ocean_lib.structures.file_objects import FilesTypeFactory
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
