+++ tests/readmes/test_dispenser_flow.py
@@ -6,10 +6,10 @@ import os
 
 from ocean_lib.example_config import ExampleConfig
 from ocean_lib.models.dispenser import Dispenser
-from ocean_lib.models.models_structures import CreateErc20Data, DispenserData
 from ocean_lib.ocean.mint_fake_ocean import mint_fake_OCEAN
 from ocean_lib.ocean.ocean import Ocean
 from ocean_lib.ocean.util import get_address_of_type
+from ocean_lib.structures.abi_tuples import CreateErc20Data, DispenserData
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.wallet import Wallet
